rename: demo plugin package name online.aruka.demo to online.aruka.custom_crafter
