+++ demo/src/main/kotlin/online/aruka/custom_crafter/listener/CustomItemRegisteredListener.kt
@@ -1,7 +1,7 @@
-package online.aruka.demo.listener
+package online.aruka.custom_crafter.listener
 
 import io.github.sakaki_aruka.customcrafter.event.RegisterCustomRecipeEvent
-import online.aruka.demo.Demo
+import online.aruka.custom_crafter.Demo
 import org.bukkit.event.EventHandler
 import org.bukkit.event.Listener
