+++ demo/src/main/resources/plugin.yml
@@ -1,6 +1,6 @@
 name: demo
 version: '${project.version}' #'5.0.15'
-main: online.aruka.demo.Demo
+main: online.aruka.custom_crafter.Demo
 api-version: '${paper_api.version}' #'1.21'
 depend: ["Custom_Crafter"]
 authors: ["Sakaki-Aruka"]
